feat: UIAlertController convenience init 추가,  삭제 버튼 클릭시 알람 표시 뜨도록 추가 #29

+++ Coin/TradingLog/TradingLogViewController.swift
@@ -74,8 +74,10 @@ extension TradingLogViewController: UITableViewDelegate {
                   let tradinglogDate = row.date else {
                 return
             }
-            
-            self?.dispatch?(.deleteTradingLog(tradinglogDate))
+            let alert = UIAlertController { _ in
+                self?.dispatch?(.deleteTradingLog(tradinglogDate))
+            }
+            self?.present(alert, animated: true, completion: nil)
         }
         
         deleteAction.backgroundColor = .systemBackground
